Doc reorg and fixes
