+++ Cargo.toml
@@ -35,8 +35,9 @@ atelier-loader = { version = "0.1", path = "loader" }
 [features]
 serde-1 = ["atelier-core/serde"]
 asset_uuid_macro = ["atelier-core/asset_uuid_macro"]
+type_uuid = ["atelier-core/type_uuid"]
 serde_importers = ["atelier-importer/serde_importers"]
 parallel_hash = [ "atelier-daemon/parallel_hash" ]
 pretty_log = [ "atelier-daemon/pretty_log" ]
-rpc_loader = ["atelier-loader/rpc_loader"]
+rpc_io = ["atelier-loader/rpc_io"]
 handle = ["atelier-loader/handle"]
