+++ core/Cargo.toml
@@ -8,12 +8,14 @@ license = "MIT OR Apache-2.0"
 [features]
 serde-1 = ["serde"]
 asset_uuid_macro = ["asset-uuid"]
+type_uuid = ["type-uuid"]
 
 [dependencies]
 uuid = { version = "0.8", features = [ "v4" ] }
 asset-uuid = { path = "./asset-uuid", version = "0.1.0", optional = true }
 serde = { version = "1.0", optional = true, features = ["derive"] }
 futures-core = { version = "0.3", default-features = false, features = ["alloc"] }
+type-uuid = { version = "0.1", optional = true, default-features = false }
 
 [dev-dependencies]
 type-uuid = "0.1"
