+++ examples/daemon_with_loader/src/image.rs
@@ -1,5 +1,7 @@
-use atelier_core::AssetUuid;
-use atelier_importer::{AsyncImporter, Error, ImportedAsset, ImporterValue, Result};
+use atelier_assets::{
+    core::AssetUuid,
+    importer::{AsyncImporter, Error, ImportedAsset, ImporterValue, Result},
+};
 use futures_core::future::BoxFuture;
 use futures_io::AsyncRead;
 use futures_util::AsyncReadExt;
