+++ examples/daemon_with_loader/src/main.rs
@@ -1,11 +1,11 @@
 mod game;
 mod image;
-use atelier_daemon::AssetDaemon;
+use atelier_assets::daemon::{init_logging, AssetDaemon};
 pub use game::Storage;
 use std::path::PathBuf;
 
 fn main() {
-    atelier_daemon::init_logging().expect("failed to init logging");
+    init_logging().expect("failed to init logging");
     std::thread::spawn(move || {
         AssetDaemon::default()
             .with_importer("png", crate::image::ImageImporter)
