+++ examples/handle_integration/Cargo.toml
@@ -7,10 +7,7 @@ readme = "README.md"
 license = "MIT/Apache-2.0"
 
 [dependencies]
-atelier-core = { version = "0.1", path = "../../core"}
-atelier-daemon = { version = "0.1", path = "../../daemon" }
-atelier-importer = { version = "0.1", path = "../../importer", features = [ "serde_importers" ] }
-atelier-loader = { version = "0.1", path = "../../loader", features = [ "asset_uuid_macro"]  }
+atelier-assets = { version = "0.1", path = "../..", features = ["serde_importers", "asset_uuid_macro"]}
 futures-executor = { version = "0.3", default-features = false }
 tokio = { version = "0.2", features = ["io-util"] }
