+++ importer/Cargo.toml
@@ -9,7 +9,6 @@ license = "MIT OR Apache-2.0"
 atelier-core = { path = "../core", version = "0.1.0", features = ["serde-1"] }
 uuid = { version = "0.8", features = ["v4"] }
 serde = "1.0"
-type-uuid = "0.1"
 erased-serde = "0.3"
 ron = { version = "0.5", optional = true }
 typetag = { version = "0.1", optional = true }
@@ -18,9 +17,10 @@ futures-io = { version = "0.3", default-features = false }
 futures-lite = { version = "1.3" }
 futures-core = { version = "0.3", default-features = false, features = ["alloc"] }
 log = { version = "0.4", features = ["serde"] }
+type-uuid = { version = "0.1", optional = true }
 
 [dev-dependencies]
 futures-executor = { version = "0.3" }
 
 [features]
-serde_importers = ["typetag", "serde-importable-derive", "ron"]
+serde_importers = ["typetag", "serde-importable-derive", "ron", "atelier-core/type_uuid", "type-uuid"]
