+++ importer/src/ron_importer.rs
@@ -3,7 +3,7 @@ use atelier_core::AssetUuid;
 use ron::de::from_reader;
 use serde::{Deserialize, Serialize};
 use std::io::Read;
-use type_uuid::TypeUuid;
+use type_uuid::*;
 
 #[derive(Default, Deserialize, Serialize, TypeUuid, Clone, Copy)]
 #[uuid = "f3cd048a-2c98-4e4b-95a2-d7c0ee6f7beb"]
